+++ packages/types/NProfilePointer.ts
@@ -2,6 +2,6 @@
 export interface NProfilePointer {
   /** Profile public key in hex format. */
   pubkey: string;
-  /** Relays in which the profile is more likely to be found. */
-  relays?: WebSocket['url'][];
+  /** Relays URLs in which the profile is more likely to be found. */
+  relays?: string[];
 }